+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [FEAT] Implement  `opentracing`
 - [REFACTOR] Port `infra/sarama` to `pkg`
 - [REFACTOR] Update ChainID manipulation in base 10 decimal format
+- [FIX] Remove `Keys` on `engine.TxContext`
 
 ### Version 0.3.4